requirements.txt
SpeechRecognition
pydub
requests
azure-cognitiveservices-speech